Move resolve helper into a module and extend World to enable grabbing to work
